Data.Type.Index.Trans

Documentation

data IxComp (i :: k -> l -> *) (j :: l -> m -> *) :: k -> m -> * where

Constructors

IxComp :: !(i a b) -> !(j b c) -> IxComp i j a c

data IxJust (i :: k -> l -> *) :: Maybe k -> l -> * where

Constructors

IxJust :: !(i a b) -> IxJust i (Just a) b

data IxOr (i :: k -> m -> *) (j :: l -> m -> *) :: Either k l -> m -> * where

Constructors

IxOrL :: !(i a b) -> IxOr i j (Left a) b

IxOrR :: !(j a b) -> IxOr i j (Right a) b

data IxSecond (i :: k -> l -> *) :: (m, k) -> l -> * where

Constructors

IxSecond :: !(i a b) -> IxSecond i '(c, a) b

data IxFirst (i :: k -> l -> *) :: (k, m) -> l -> * where

Constructors

IxFirst :: !(i a b) -> IxFirst i '(a, c) b

data IxList (i :: k -> l -> *) :: [k] -> l -> * where

Constructors

IxHead :: !(i a b) -> IxList i (a :< as) b

IxTail :: !(IxList i as b) -> IxList i (a :< as) b

class IxLift (t :: (k -> m -> *) -> l -> m -> *) (x :: l) where

Minimal complete definition

ixLift

Associated Types

type LiftI t x :: k

Methods

ixLift :: i (LiftI t x) y -> t i x y

type IxEnv = IxList (IxFirst (:~:))

type IxList' = IxList (:~:)

data (a :: k) :~: (b :: k) :: forall k. k -> k -> * where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: a :~: a