Data.Type.Quantifier

Description

Types for working with (and under) existentially and universally quantified types.

The Some type can be very useful when working with type-indexed GADTs, where defining instances for classes like Read can be tedious at best, and frequently impossible, for the GADT itself.

Documentation

data Some f :: * where

Constructors

Some :: f a -> Some f

some :: Some f -> (forall a. f a -> r) -> r

An eliminator for a Some type.

Consider this function akin to a Monadic bind, except instead of binding into a Monad with a sequent function, we're binding into the existential quantification with a universal eliminator function.

It serves as an explicit delimiter in a program of where the type index may be used and depended on, and where it may not.

NB: the result type of the eliminating function may not refer to the universally quantified type index a.

(>>-) :: Some f -> (forall a. f a -> r) -> r infixl 1

(>->) :: (forall x. f x -> Some g) -> (forall x. g x -> Some h) -> f a -> Some h infixr 1

withSome :: (forall a. f a -> r) -> Some f -> r

onSome :: (forall a. f a -> g x) -> Some f -> Some g

msome :: Monad m => f a -> m (Some f)

(>>=-) :: Monad m => m (Some f) -> (forall a. f a -> m r) -> m r infixl 1

data Some2 f :: * where

Constructors

Some2 :: f a b -> Some2 f

some2 :: Some2 f -> (forall a b. f a b -> r) -> r

(>>--) :: Some2 f -> (forall a b. f a b -> r) -> r infixl 1

(>-->) :: (forall x y. f x y -> Some2 g) -> (forall x y. g x y -> Some2 h) -> f a b -> Some2 h infixr 1

withSome2 :: (forall a b. f a b -> r) -> Some2 f -> r

onSome2 :: (forall a b. f a b -> g x y) -> Some2 f -> Some2 g

msome2 :: Monad m => f a b -> m (Some2 f)

(>>=--) :: Monad m => m (Some2 f) -> (forall a b. f a b -> m r) -> m r infixl 1

data Some3 f :: * where

Constructors

Some3 :: f a b c -> Some3 f

some3 :: Some3 f -> (forall a b c. f a b c -> r) -> r

(>>---) :: Some3 f -> (forall a b c. f a b c -> r) -> r infixl 1

(>--->) :: (forall x y z. f x y z -> Some3 g) -> (forall x y z. g x y z -> Some3 h) -> f a b c -> Some3 h infixr 1

withSome3 :: (forall a b c. f a b c -> r) -> Some3 f -> r

onSome3 :: (forall a b c. f a b c -> g x y z) -> Some3 f -> Some3 g

msome3 :: Monad m => f a b c -> m (Some3 f)

(>>=---) :: Monad m => m (Some3 f) -> (forall a b c. f a b c -> m r) -> m r infixl 1

data SomeC c f where

Constructors

SomeC :: c a => f a -> SomeC c f

someC :: SomeC c f -> (forall a. c a => f a -> r) -> r

(>>~) :: SomeC c f -> (forall a. c a => f a -> r) -> r infixl 1

msomeC :: (Monad m, c a) => f a -> m (SomeC c f)

(>>=~) :: Monad m => m (SomeC c f) -> (forall a. c a => f a -> m r) -> m r infixl 1

data Every f :: * where

Constructors

Every :: (forall a. f a) -> Every f

Fields

instEvery :: forall a. f a

data Every2 f :: * where

Constructors

Every2 :: (forall a b. f a b) -> Every2 f

Fields

instEvery2 :: forall a b. f a b

data Every3 f :: * where

Constructors

Every3 :: (forall a b c. f a b c) -> Every3 f

Fields

instEvery3 :: forall a b c. f a b c

data EveryC c f :: * where

Constructors

EveryC :: (forall a. c a => f a) -> EveryC c f

Fields

instEveryC :: forall a. c a => f a