Data.Type.Sum

Description

Sum is a type combinators for representing disjoint sums of indices (as :: [k]) of a single functor @(f :: k -> *). Contrast to the many-functors-one-index FSum

Documentation

data Sum f :: [k] -> * where

Constructors

InL :: !(f a) -> Sum f (a :< as)

InR :: !(Sum f as) -> Sum f (a :< as)

nilSum :: Sum f Ø -> Void

There are no possible values of the type Sum f Ø.

decomp :: Sum f (a :< as) -> Either (f a) (Sum f as)

injectSum :: Index as a -> f a -> Sum f as

inj :: a ∈ as => f a -> Sum f as

prj :: a ∈ as => Sum f as -> Maybe (f a)

index :: Index as a -> Sum f as -> Maybe (f a)

elimSum :: (forall x xs. f x -> p (x :< xs)) -> (forall x xs. Index as x -> p xs -> p (x :< xs)) -> Sum f as -> p as