Type.Class.Higher

Description

Higher order analogs of type classes from the Prelude.

Documentation

class Eq1 f where

Minimal complete definition

Nothing

Methods

eq1 :: f a -> f a -> Bool

neq1 :: f a -> f a -> Bool

(=#=) :: Eq1 f => f a -> f a -> Bool infix 4

class Eq2 f where

Minimal complete definition

Nothing

Methods

eq2 :: f a b -> f a b -> Bool

neq2 :: f a b -> f a b -> Bool

(=##=) :: Eq2 f => f a b -> f a b -> Bool infix 4

class Eq3 f where

Minimal complete definition

Nothing

Methods

eq3 :: f a b c -> f a b c -> Bool

neq3 :: f a b c -> f a b c -> Bool

(=###=) :: Eq3 f => f a b c -> f a b c -> Bool infix 4

class Eq1 f => Ord1 f where

Minimal complete definition

Nothing

Methods

compare1 :: f a -> f a -> Ordering

(<#) :: f a -> f a -> Bool infix 4

(>#) :: f a -> f a -> Bool infix 4

(<=#) :: f a -> f a -> Bool infix 4

(>=#) :: f a -> f a -> Bool infix 4

class Eq2 f => Ord2 f where

Minimal complete definition

Nothing

Methods

compare2 :: f a b -> f a b -> Ordering

(<##) :: f a b -> f a b -> Bool infix 4

(>##) :: f a b -> f a b -> Bool infix 4

(<=##) :: f a b -> f a b -> Bool infix 4

(>=##) :: f a b -> f a b -> Bool infix 4

class Eq3 f => Ord3 f where

Minimal complete definition

Nothing

Methods

compare3 :: f a b c -> f a b c -> Ordering

(<###) :: f a b c -> f a b c -> Bool infix 4

(>###) :: f a b c -> f a b c -> Bool infix 4

(<=###) :: f a b c -> f a b c -> Bool infix 4

(>=###) :: f a b c -> f a b c -> Bool infix 4

class Show1 f where

Minimal complete definition

Nothing

Methods

showsPrec1 :: Int -> f a -> ShowS

show1 :: f a -> String

shows1 :: Show1 f => f a -> ShowS

class Show2 f where

Minimal complete definition

Nothing

Methods

showsPrec2 :: Int -> f a b -> ShowS

show2 :: f a b -> String

shows2 :: Show2 f => f a b -> ShowS

class Show3 f where

Minimal complete definition

Nothing

Methods

showsPrec3 :: Int -> f a b c -> ShowS

show3 :: f a b c -> String

shows3 :: Show3 f => f a b c -> ShowS

class Read1 f where

Methods

readsPrec1 :: Int -> ReadS (Some f)

reads1 :: Read1 f => ReadS (Some f)

readMaybe1 :: Read1 f => String -> Maybe (Some f)

class Read2 f where

Methods

readsPrec2 :: Int -> ReadS (Some2 f)

reads2 :: Read2 f => ReadS (Some2 f)

readMaybe2 :: Read2 f => String -> Maybe (Some2 f)

class Read3 f where

Methods

readsPrec3 :: Int -> ReadS (Some3 f)

reads3 :: Read3 f => ReadS (Some3 f)

readMaybe3 :: Read3 f => String -> Maybe (Some3 f)

class Functor1 t where

Methods

map1 :: (forall a. f a -> g a) -> t f b -> t g b

Take a natural transformation to a lifted natural transformation.

class IxFunctor1 i t | t -> i where

Methods

imap1 :: (forall a. i b a -> f a -> g a) -> t f b -> t g b

class Foldable1 t where

Methods

foldMap1 :: Monoid m => (forall a. f a -> m) -> t f b -> m

class IxFoldable1 i t | t -> i where

Methods

ifoldMap1 :: Monoid m => (forall a. i b a -> f a -> m) -> t f b -> m

class (Functor1 t, Foldable1 t) => Traversable1 t where

Methods

traverse1 :: Applicative h => (forall a. f a -> h (g a)) -> t f b -> h (t g b)

class (IxFunctor1 i t, IxFoldable1 i t) => IxTraversable1 i t | t -> i where

Methods

itraverse1 :: Applicative h => (forall a. i b a -> f a -> h (g a)) -> t f b -> h (t g b)

class Bifunctor1 t where

Methods

bimap1 :: (forall a. f a -> h a) -> (forall a. g a -> i a) -> t f g b -> t h i b

class IxBifunctor1 i j t | t -> i j where

Methods

ibimap1 :: (forall a. i b a -> f a -> f' a) -> (forall a. j b a -> g a -> g' a) -> t f g b -> t f' g' b