Type.Class.Witness

Description

A type t that is a Witness p q t provides a Constraint entailment of q, given that p holds.

The Witness class uses an associated Constraint WitnessC to maintain backwards inference of Witness instances with respect to type refinement. See the Known class for more information.

Heavily inspired by ekmett's constraints library: http://hackage.haskell.org/package/constraints

The code provided here does not quite subsume the constraints library, as we do not give classes and instances for representing the standard library's class heirarchy and instance definitions.

Documentation

data Wit :: Constraint -> * where

A reified Constraint.

Constructors

Wit :: c => Wit c

data Wit1 :: (k -> Constraint) -> k -> * where

Constructors

Wit1 :: c a => Wit1 c a

data (:-) :: Constraint -> Constraint -> * where infixr 4

Reified evidence of Constraint entailment.

Given a term of p :- q, the Constraint q holds if p holds.

Entailment of Constraints form a Category:

>>> id  :: p :- p
>>> (.) :: (q :- r) -> (p :-> q) -> (p :- r)

Constructors

Sub :: (p => Wit q) -> p :- q

Fields

getSub :: p => Wit q

transC :: (b :- c) -> (a :- b) -> a :- c

class WitnessC p q t => Witness p q t | t -> p q where

A general eliminator for entailment.

Given a term of type t with an instance Witness p q t and a term of type r that depends on Constraint q, we can reduce the Constraint to p.

If p is ØC, i.e. the empty Constraint (), then a Witness t can completely discharge the Constraint q.

Associated Types

type WitnessC p q t :: Constraint

Methods

(\\) :: p => (q => r) -> t -> r infixl 1

(//) :: (Witness p q t, p) => t -> (q => r) -> r infixr 0

witnessed :: Witness ØC q t => t -> Wit q

Convert a Witness to a canonical reified Constraint.

entailed :: Witness p q t => t -> p :- q

Convert a Witness to a canonical reified entailment.

class Fails c where

Methods

failC :: c :- Fail

absurdC :: Fails a => a :- b

class c => Const c d where

Methods

constC :: Wit c

class f (g a) => (f ∘ g) a where infixr 9

Methods

compC :: Wit (f (g a))

class (f a, g a) => (f ∧ g) a where infixr 7

Methods

conjC :: (Wit (f a), Wit (g a))

class (f ∨ g) a where infixr 6

Methods

disjC :: Either (Wit (f a)) (Wit (g a))

eitherC :: forall f g a b. (f a :- b) -> (g a :- b) -> (f ∨ g) a :- b

pureC :: b => a :- b

contraC :: (a :- Fail) -> a :- b

class Forall p q where

Minimal complete definition

Nothing

Methods

forall :: p a :- q a

toEquality :: ((a ~ b) :- (c ~ d)) -> (a :~: b) -> c :~: d

commute :: (a ~ b) :- (b ~ a)

type family Holds b c :: Constraint

Equations

Holds True c = c

Holds False c = ØC

falso :: (b ~ False) :- Holds b c

top :: a :- ØC

bottom :: Fail :- c

(//?) :: (Witness p q t, p) => Maybe t -> (q => Maybe r) -> Maybe r infixr 0

Constraint chaining under Maybe.

(//?+) :: (Witness p q t, p) => Either e t -> (q => Either e r) -> Either e r infixr 0

witMaybe :: (Witness p q t, p) => Maybe t -> (q => Maybe r) -> Maybe r -> Maybe r

qed :: Maybe (a :~: a)

impossible :: a -> Void

exFalso :: Wit Fail -> a

(=?=) :: TestEquality f => f a -> f b -> Maybe (a :~: b) infix 4

class TestEquality1 f where

Methods

testEquality1 :: f a c -> f b c -> Maybe (a :~: b)

(=??=) :: TestEquality1 f => f a c -> f b c -> Maybe (a :~: b) infix 4

data Dec a

Constructors

Proven a

Refuted (a -> Void)

class DecEquality f where

Methods

decideEquality :: f a -> f b -> Dec (a :~: b)

decCase :: Dec a -> (a -> r) -> ((a -> Void) -> r) -> r

absurd :: Arrow p => p Void a