Type.Family.Nat

Description

Type-level natural numbers, along with frequently used type families over them.

Documentation

data N

Constructors

Z

S N

fromInt :: Int -> Maybe N

type family IsZero (x :: N) :: Bool where ...

Equations

IsZero Z = True

IsZero (S x) = False

zeroCong :: (x ~ y) :- (IsZero x ~ IsZero y)

zNotS :: (Z ~ S x) :- Fail

type NatEq (x :: N) (y :: N) = x == y

Note: Was previously a type family, but is now just a synonym for == from Equality.

type family Iota (x :: N) = (xs :: [N]) | xs -> x where ...

Equations

Iota Z = Ø

Iota (S x) = x :< Iota x

iotaCong :: (x ~ y) :- (Iota x ~ Iota y)

type family Pred (x :: N) :: N where ...

Equations

Pred (S n) = n

type Pos n = n ~ S (Pred n)

predCong :: (x ~ y) :- (Pred x ~ Pred y)

type family (x :: N) + (y :: N) :: N where ... infixr 6

Equations

Z + y = y

(S x) + y = S (x + y)

data AddW (f :: N -> *) :: N -> * where

Constructors

AddW :: !(f x) -> !(f y) -> AddW f (x + y)

addCong :: (w ~ y, x ~ z) :- ((w + x) ~ (y + z))

type family (x :: N) * (y :: N) :: N where ... infixr 7

Equations

Z * y = Z

(S x) * y = (x * y) + y

data MulW (f :: N -> *) :: N -> * where

Constructors

MulW :: !(f x) -> !(f y) -> MulW f (x * y)

mulCong :: (w ~ y, x ~ z) :- ((w * x) ~ (y * z))

type family (x :: N) ^ (y :: N) :: N where ... infixl 8

Equations

x ^ Z = S Z

x ^ (S y) = (x ^ y) * x

expCong :: (w ~ y, x ~ z) :- ((w ^ x) ~ (y ^ z))

type family Len (as :: [k]) :: N where ...

Equations

Len Ø = Z

Len (a :< as) = S (Len as)

lenCong :: (as ~ bs) :- (Len as ~ Len bs)

type family Ix (x :: N) (as :: [k]) :: k where ...

Equations

Ix Z (a :< as) = a

Ix (S x) (a :< as) = Ix x as

ixCong :: (x ~ y, as ~ bs) :- (Ix x as ~ Ix y bs)

type family (x :: N) < (y :: N) :: Bool where ... infix 4

Equations

Z < Z = False

Z < (S y) = True

(S x) < Z = False

(S x) < (S y) = x < y

type (<=) x y = (x == y) || (x < y) infix 4

type family (x :: N) > (y :: N) :: Bool where ... infix 4

Equations

Z > Z = False

Z > (S y) = False

(S x) > Z = True

(S x) > (S y) = x > y

type (>=) x y = (x == y) || (x > y) infix 4

type N0 = Z

Convenient aliases for low-value Peano numbers.

type N1 = S N0

type N2 = S N1

type N3 = S N2

type N4 = S N3

type N5 = S N4

type N6 = S N5

type N7 = S N6

type N8 = S N7

type N9 = S N8

type N10 = S N9