Data.Type.Vector

Description

Vec and its combinator analog VecT represent lists of known length, characterized by the index (n :: N) in Vec n a or VecT n f a.

The classic example used ad nauseum for type-level programming.

The operations on Vec and VecT correspond to the type level arithmetic operations on the kind N.

Documentation

data VecT (n :: N) (f :: k -> *) :: k -> * where

Constructors

ØV :: VecT Z f a

(:*) :: !(f a) -> !(VecT n f a) -> VecT (S n) f a infixr 4

(*:) :: f a -> f a -> VecT (S (S Z)) f a infix 5

elimVecT :: p Z -> (forall x. f a -> p x -> p (S x)) -> VecT n f a -> p n

elimV :: p Z -> (forall x. a -> p x -> p (S x)) -> Vec n a -> p n

type Vec n = VecT n I

pattern (:+) :: a -> Vec n a -> Vec (S n) a infixr 4

(+:) :: a -> a -> Vec (S (S Z)) a infix 5

(.++) :: VecT x f a -> VecT y f a -> VecT (x + y) f a infixr 5

vrep :: forall n f a. Known Nat n => f a -> VecT n f a

head' :: VecT (S n) f a -> f a

tail' :: VecT (S n) f a -> VecT n f a

onTail :: (VecT m f a -> VecT n f a) -> VecT (S m) f a -> VecT (S n) f a

vDel :: Fin n -> VecT n f a -> VecT (Pred n) f a

imap :: (Fin n -> f a -> g b) -> VecT n f a -> VecT n g b

ifoldMap :: Monoid m => (Fin n -> f a -> m) -> VecT n f a -> m

itraverse :: Applicative h => (Fin n -> f a -> h (g b)) -> VecT n f a -> h (VecT n g b)

index :: Fin n -> VecT n f a -> f a

index' :: Fin n -> Vec n a -> a

vmap :: (f a -> g b) -> VecT n f a -> VecT n g b

vap :: (f a -> g b -> h c) -> VecT n f a -> VecT n g b -> VecT n h c

vfoldr :: (f a -> b -> b) -> b -> VecT n f a -> b

vfoldMap' :: (b -> b -> b) -> b -> (f a -> b) -> VecT n f a -> b

vfoldMap :: Monoid m => (f a -> m) -> VecT n f a -> m

withVecT :: [f a] -> (forall n. VecT n f a -> r) -> r

withV :: [a] -> (forall n. Vec n a -> r) -> r

findV :: Eq a => a -> Vec n a -> Maybe (Fin n)

findVecT :: Eq (f a) => f a -> VecT n f a -> Maybe (Fin n)

newtype M ns a

Constructors

M

Fields

getMatrix :: Matrix ns a

type family Matrix (ns :: [N]) :: * -> * where ...

Equations

Matrix Ø = I

Matrix (n :< ns) = VecT n (Matrix ns)

vgen_ :: Known Nat n => (Fin n -> f a) -> VecT n f a

vgen :: Nat n -> (Fin n -> f a) -> VecT n f a

mgen_ :: Known (Prod Nat) ns => (Prod Fin ns -> a) -> M ns a

mgen :: Prod Nat ns -> (Prod Fin ns -> a) -> M ns a

onMatrix :: (Matrix ms a -> Matrix ns b) -> M ms a -> M ns b

diagonal :: VecT n (VecT n f) a -> VecT n f a

vtranspose :: Known Nat n => VecT m (VecT n f) a -> VecT n (VecT m f) a

transpose :: Known Nat n => M (m :< (n :< ns)) a -> M (n :< (m :< ns)) a

m0 :: M Ø Int

m1 :: M '[N2] Int

m2 :: M '[N2, N4] Int

m3 :: M '[N2, N3, N4] (Int, Int, Int)

m4 :: M '[N2, N3, N4, N5] (Int, Int, Int, Int)

ppVec :: (VecT n ((->) String) String -> ShowS) -> (f a -> ShowS) -> VecT n f a -> ShowS

ppMatrix :: forall ns a. (Show a, Known Length ns) => M ns a -> IO ()

ppMatrix' :: Show a => Length ns -> Matrix ns a -> ShowS

mzipWith :: Monoid a => (a -> a -> b) -> [a] -> [a] -> [b]

zipLines :: (ShowS -> ShowS -> ShowS) -> ShowS -> ShowS -> ShowS

compose :: Foldable f => f (a -> a) -> a -> a