Type.Class.Higher

Description

Higher order analogs of type classes from the Prelude, and quantifier data types.

Documentation

class Eq1 (f :: k -> *) where

Methods

eq1 :: f a -> f a -> Bool

eq1 :: Eq (f a) => f a -> f a -> Bool

neq1 :: f a -> f a -> Bool

(=#=) :: Eq1 f => f a -> f a -> Bool infix 4

class Eq2 (f :: k -> l -> *) where

Methods

eq2 :: f a b -> f a b -> Bool

eq2 :: Eq (f a b) => f a b -> f a b -> Bool

neq2 :: f a b -> f a b -> Bool

(=##=) :: Eq2 f => f a b -> f a b -> Bool infix 4

class Eq3 (f :: k -> l -> m -> *) where

Methods

eq3 :: f a b c -> f a b c -> Bool

eq3 :: Eq (f a b c) => f a b c -> f a b c -> Bool

neq3 :: f a b c -> f a b c -> Bool

(=###=) :: Eq3 f => f a b c -> f a b c -> Bool infix 4

class Eq1 f => Ord1 (f :: k -> *) where

Methods

compare1 :: f a -> f a -> Ordering

compare1 :: Ord (f a) => f a -> f a -> Ordering

(<#) :: f a -> f a -> Bool infix 4

(>#) :: f a -> f a -> Bool infix 4

(<=#) :: f a -> f a -> Bool infix 4

(>=#) :: f a -> f a -> Bool infix 4

class Eq2 f => Ord2 (f :: k -> l -> *) where

Methods

compare2 :: f a b -> f a b -> Ordering

compare2 :: Ord (f a b) => f a b -> f a b -> Ordering

(<##) :: f a b -> f a b -> Bool infix 4

(>##) :: f a b -> f a b -> Bool infix 4

(<=##) :: f a b -> f a b -> Bool infix 4

(>=##) :: f a b -> f a b -> Bool infix 4

class Eq3 f => Ord3 (f :: k -> l -> m -> *) where

Methods

compare3 :: f a b c -> f a b c -> Ordering

compare3 :: Ord (f a b c) => f a b c -> f a b c -> Ordering

(<###) :: f a b c -> f a b c -> Bool infix 4

(>###) :: f a b c -> f a b c -> Bool infix 4

(<=###) :: f a b c -> f a b c -> Bool infix 4

(>=###) :: f a b c -> f a b c -> Bool infix 4

class Show1 (f :: k -> *) where

Methods

showsPrec1 :: Int -> f a -> ShowS

showsPrec1 :: Show (f a) => Int -> f a -> ShowS

show1 :: f a -> String

shows1 :: Show1 f => f a -> ShowS

class Show2 (f :: k -> l -> *) where

Methods

showsPrec2 :: Int -> f a b -> ShowS

showsPrec2 :: Show (f a b) => Int -> f a b -> ShowS

show2 :: f a b -> String

shows2 :: Show2 f => f a b -> ShowS

class Show3 (f :: k -> l -> m -> *) where

Methods

showsPrec3 :: Int -> f a b c -> ShowS

showsPrec3 :: Show (f a b c) => Int -> f a b c -> ShowS

show3 :: f a b c -> String

shows3 :: Show3 f => f a b c -> ShowS

class Read1 (f :: k -> *) where

Minimal complete definition

readsPrec1

Methods

readsPrec1 :: Int -> ReadS (Some f)

reads1 :: Read1 f => ReadS (Some f)

readMaybe1 :: Read1 f => String -> Maybe (Some f)

class Read2 (f :: k -> l -> *) where

Minimal complete definition

readsPrec2

Methods

readsPrec2 :: Int -> ReadS (Some2 f)

reads2 :: Read2 f => ReadS (Some2 f)

readMaybe2 :: Read2 f => String -> Maybe (Some2 f)

class Read3 (f :: k -> l -> m -> *) where

Minimal complete definition

readsPrec3

Methods

readsPrec3 :: Int -> ReadS (Some3 f)

reads3 :: Read3 f => ReadS (Some3 f)

readMaybe3 :: Read3 f => String -> Maybe (Some3 f)

class Functor1 (t :: (k -> *) -> l -> *) where

Minimal complete definition

map1

Methods

map1 :: (forall (a :: k). f a -> g a) -> t f b -> t g b

Take a natural transformation to a lifted natural transformation.

class IxFunctor1 (i :: l -> k -> *) (t :: (k -> *) -> l -> *) | t -> i where

Minimal complete definition

imap1

Methods

imap1 :: (forall (a :: k). i b a -> f a -> g a) -> t f b -> t g b

class Foldable1 (t :: (k -> *) -> l -> *) where

Minimal complete definition

foldMap1

Methods

foldMap1 :: Monoid m => (forall (a :: k). f a -> m) -> t f b -> m

class IxFoldable1 (i :: l -> k -> *) (t :: (k -> *) -> l -> *) | t -> i where

Minimal complete definition

ifoldMap1

Methods

ifoldMap1 :: Monoid m => (forall (a :: k). i b a -> f a -> m) -> t f b -> m

class (Functor1 t, Foldable1 t) => Traversable1 (t :: (k -> *) -> l -> *) where

Minimal complete definition

traverse1

Methods

traverse1 :: Applicative h => (forall (a :: k). f a -> h (g a)) -> t f b -> h (t g b)

class (IxFunctor1 i t, IxFoldable1 i t) => IxTraversable1 (i :: l -> k -> *) (t :: (k -> *) -> l -> *) | t -> i where

Minimal complete definition

itraverse1

Methods

itraverse1 :: Applicative h => (forall (a :: k). i b a -> f a -> h (g a)) -> t f b -> h (t g b)

class Bifunctor1 (t :: (k -> *) -> (l -> *) -> m -> *) where

Minimal complete definition

bimap1

Methods

bimap1 :: (forall (a :: k). f a -> h a) -> (forall (a :: l). g a -> i a) -> t f g b -> t h i b

class IxBifunctor1 (i :: m -> k -> *) (j :: m -> l -> *) (t :: (k -> *) -> (l -> *) -> m -> *) | t -> i j where

Minimal complete definition

ibimap1

Methods

ibimap1 :: (forall (a :: k). i b a -> f a -> f' a) -> (forall (a :: l). j b a -> g a -> g' a) -> t f g b -> t f' g' b

data Some (f :: k -> *) :: * where

Constructors

Some :: f a -> Some f

some :: Some f -> (forall a. f a -> r) -> r

An eliminator for a Some type.

Consider this function akin to a Monadic bind, except instead of binding into a Monad with a sequent function, we're binding into the existential quantification with a universal eliminator function.

It serves as an explicit delimiter in a program of where the type index may be used and depended on, and where it may not.

NB: the result type of the eliminating function may not refer to the universally quantified type index a.

(>>-) :: Some f -> (forall a. f a -> r) -> r infixl 1

(>->) :: (forall x. f x -> Some g) -> (forall x. g x -> Some h) -> f a -> Some h infixr 1

withSome :: (forall a. f a -> r) -> Some f -> r

onSome :: (forall a. f a -> g x) -> Some f -> Some g

msome :: Monad m => f a -> m (Some f)

(>>=-) :: Monad m => m (Some f) -> (forall a. f a -> m r) -> m r infixl 1

data Some2 (f :: k -> l -> *) :: * where

Constructors

Some2 :: f a b -> Some2 f

some2 :: Some2 f -> (forall a b. f a b -> r) -> r

(>>--) :: Some2 f -> (forall a b. f a b -> r) -> r infixl 1

(>-->) :: (forall x y. f x y -> Some2 g) -> (forall x y. g x y -> Some2 h) -> f a b -> Some2 h infixr 1

withSome2 :: (forall a b. f a b -> r) -> Some2 f -> r

onSome2 :: (forall a b. f a b -> g x y) -> Some2 f -> Some2 g

msome2 :: Monad m => f a b -> m (Some2 f)

(>>=--) :: Monad m => m (Some2 f) -> (forall a b. f a b -> m r) -> m r infixl 1

data Some3 (f :: k -> l -> m -> *) :: * where

Constructors

Some3 :: f a b c -> Some3 f

some3 :: Some3 f -> (forall a b c. f a b c -> r) -> r

(>>---) :: Some3 f -> (forall a b c. f a b c -> r) -> r infixl 1

(>--->) :: (forall x y z. f x y z -> Some3 g) -> (forall x y z. g x y z -> Some3 h) -> f a b c -> Some3 h infixr 1

withSome3 :: (forall a b c. f a b c -> r) -> Some3 f -> r

onSome3 :: (forall a b c. f a b c -> g x y z) -> Some3 f -> Some3 g

msome3 :: Monad m => f a b c -> m (Some3 f)

(>>=---) :: Monad m => m (Some3 f) -> (forall a b c. f a b c -> m r) -> m r infixl 1

data SomeC (c :: k -> Constraint) (f :: k -> *) where

Constructors

SomeC :: c a => f a -> SomeC c f

someC :: SomeC c f -> (forall a. c a => f a -> r) -> r

(>>~) :: SomeC c f -> (forall a. c a => f a -> r) -> r infixl 1

msomeC :: (Monad m, c a) => f a -> m (SomeC c f)

(>>=~) :: Monad m => m (SomeC c f) -> (forall a. c a => f a -> m r) -> m r infixl 1