Type.Class.Known

Description

The Known class, among others in this library, use an associated Constraint to maintain a bidirectional chain of inference.

For instance, given evidence of Known Nat n, if n later gets refined to n', we can correctly infer Known Nat n', as per the type instance defined for KnownC Nat (S n').

Documentation

class KnownC f a => Known (f :: k -> *) (a :: k) where

Each instance of Known provides a canonical construction of a type at a particular index.

Useful for working with singleton-esque GADTs.

Minimal complete definition

known

Associated Types

type KnownC f a :: Constraint

Methods

known :: f a