Type.Class.Witness

Description

A type t that is a Witness p q t provides a Constraint entailment of q, given that p holds.

The Witness class uses an associated Constraint WitnessC to maintain backwards inference of Witness instances with respect to type refinement. See the Known class for more information.

Heavily inspired by ekmett's constraints library: http://hackage.haskell.org/package/constraints

The code provided here does not quite subsume the constraints library, as we do not give classes and instances for representing the standard library's class heirarchy and instance definitions.

Documentation

class DecEquality (f :: k -> *) where

Minimal complete definition

decideEquality

Methods

decideEquality :: f a -> f b -> Dec (a :~: b)

data Dec a

Constructors

Proven a

Refuted (a -> Void)

class TestEquality1 (f :: k -> l -> *) where

Minimal complete definition

testEquality1

Methods

testEquality1 :: f a c -> f b c -> Maybe (a :~: b)

type family Holds (b :: Bool) (c :: Constraint) :: Constraint where ...

Equations

Holds True c = c

Holds False c = ØC

class Forall (p :: k -> Constraint) (q :: k -> Constraint) where

Methods

forall :: p a :- q a

forall :: q a => p a :- q a

class ((f :: k -> Constraint) ∨ (g :: k -> Constraint)) (a :: k) where infixr 6

Minimal complete definition

disjC

Methods

disjC :: Either (Wit (f a)) (Wit (g a))

class (f a, g a) => ((f :: k -> Constraint) ∧ (g :: k -> Constraint)) (a :: k) where infixr 7

Minimal complete definition

conjC

Methods

conjC :: (Wit (f a), Wit (g a))

class f (g a) => ((f :: l -> Constraint) ∘ (g :: k -> l)) (a :: k) where infixr 9

Minimal complete definition

compC

Methods

compC :: Wit (f (g a))

class c => Const (c :: Constraint) (d :: k) where

Minimal complete definition

constC

Methods

constC :: Wit c

class Fails (c :: Constraint) where

Minimal complete definition

failC

Methods

failC :: c :- Fail

class WitnessC p q t => Witness (p :: Constraint) (q :: Constraint) (t :: *) | t -> p q where

A general eliminator for entailment.

Given a term of type t with an instance Witness p q t and a term of type r that depends on Constraint q, we can reduce the Constraint to p.

If p is ØC, i.e. the empty Constraint (), then a Witness t can completely discharge the Constraint q.

Minimal complete definition

(\\)

Associated Types

type WitnessC p q t :: Constraint

Methods

(\\) :: p => (q => r) -> t -> r infixl 1

data (:-) :: Constraint -> Constraint -> * where infixr 4

Reified evidence of Constraint entailment.

Given a term of p :- q, the Constraint q holds if p holds.

Entailment of Constraints form a Category:

>>> id  :: p :- p
>>> (.) :: (q :- r) -> (p :-> q) -> (p :- r)

Constructors

Sub :: {..} -> p :- q

Fields

getSub :: p => Wit q

data Wit1 :: (k -> Constraint) -> k -> * where

Constructors

Wit1 :: c a => Wit1 c a

data Wit :: Constraint -> * where

A reified Constraint.

Constructors

Wit :: c => Wit c

transC :: (b :- c) -> (a :- b) -> a :- c

(//) :: (Witness p q t, p) => t -> (q => r) -> r infixr 0

witnessed :: Witness ØC q t => t -> Wit q

Convert a Witness to a canonical reified Constraint.

entailed :: Witness p q t => t -> p :- q

Convert a Witness to a canonical reified entailment.

absurdC :: Fails a => a :- b

eitherC :: forall f g a b. (f a :- b) -> (g a :- b) -> (f ∨ g) a :- b

pureC :: b => a :- b

contraC :: (a :- Fail) -> a :- b

toEquality :: ((a ~ b) :- (c ~ d)) -> (a :~: b) -> c :~: d

commute :: (a ~ b) :- (b ~ a)

falso :: (b ~ False) :- Holds b c

top :: a :- ØC

bottom :: Fail :- c

(//?) :: (Witness p q t, p) => Maybe t -> (q => Maybe r) -> Maybe r infixr 0

Constraint chaining under Maybe.

(//?+) :: (Witness p q t, p) => Either e t -> (q => Either e r) -> Either e r infixr 0

witMaybe :: (Witness p q t, p) => Maybe t -> (q => Maybe r) -> Maybe r -> Maybe r

qed :: Maybe (a :~: a)

impossible :: a -> Void

exFalso :: Wit Fail -> a

(=?=) :: TestEquality f => f a -> f b -> Maybe (a :~: b) infix 4

(=??=) :: TestEquality1 f => f a c -> f b c -> Maybe (a :~: b) infix 4

decCase :: Dec a -> (a -> r) -> ((a -> Void) -> r) -> r

absurd :: Arrow p => p Void a

class a ~# b => (a :: k0) ~~ (b :: k1)

Lifted, heterogeneous equality. By lifted, we mean that it can be bogus (deferred type error). By heterogeneous, the two types a and b might have different kinds. Because ~~ can appear unexpectedly in error messages to users who do not care about the difference between heterogeneous equality ~~ and homogeneous equality ~, this is printed as ~ unless -fprint-equality-relations is set.

vacuous :: Functor f => f Void -> f a

If Void is uninhabited then any Functor that holds only values of type Void is holding no values.

Since: base-4.8.0.0

data Void

Uninhabited data type

Since: base-4.8.0.0

outer :: (f a :~: g b) -> f :~: g

Extract equality of type constructors from an equality of applied types

inner :: (f a :~: g b) -> a :~: b

Extract equality of the arguments from an equality of applied types

apply :: (f :~: g) -> (a :~: b) -> f a :~: g b

Apply one equality to another, respectively

gcastWith :: (a :~: b) -> ((a ~ b) -> r) -> r

Generalized form of type-safe cast using propositional equality

castWith :: (a :~: b) -> a -> b

Type-safe cast, using propositional equality

trans :: (a :~: b) -> (b :~: c) -> a :~: c

Transitivity of equality

sym :: (a :~: b) -> b :~: a

Symmetry of equality

data (a :: k) :~: (b :: k) :: forall k. k -> k -> * where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: a :~: a

data (a :: k1) :~~: (b :: k2) :: forall k1 k2. k1 -> k2 -> * where infix 4

Kind heterogeneous propositional equality. Like :~:, a :~~: b is inhabited by a terminating value if and only if a is the same type as b.

Since: base-4.10.0.0

Constructors

HRefl :: a :~~: a

class TestEquality (f :: k -> *) where

This class contains types where you can learn the equality of two types from information contained in terms. Typically, only singleton types should inhabit this class.

Minimal complete definition

testEquality

Methods

testEquality :: f a -> f b -> Maybe (a :~: b)

Conditionally prove the equality of a and b.

type family (a :: k) == (b :: k) :: Bool where ... infix 4

A type family to compute Boolean equality.

Equations

(f a :: k2) == (g b :: k2) = (f == g) && (a == b)

(a :: k) == (a :: k) = True

(_1 :: k) == (_2 :: k) = False